Fix(CI): JPQL count 결과 타입에 맞게 followerCount를 long으로 변경

+++ semicolon/src/main/java/dukku/semicolon/shared/product/dto/follow/FollowedSellerCardResponse.java
@@ -9,6 +9,6 @@ public record FollowedSellerCardResponse(
         String intro,
         BigDecimal averageRating,
         int reviewCount,
-        int followerCount,
+        long followerCount,
         boolean followed
 ) {}
